requirements.txt
# ── Vector Storage & Embeddings (local) ──────────────────
faiss-cpu>=1.7.0
scikit-learn>=1.3.0
numpy>=1.24.0

# ── Backend API ─────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
waitress>=3.0.0

# ── LLM ─────────────────────────────────────────────────
groq>=0.4.0

# ── Utilities ───────────────────────────────────────────
python-dotenv>=1.0.0
